+++ app/src/main/java/io/github/gsantner/opoc/util/SimpleMarkdownParser.java
@@ -5,7 +5,7 @@
  * worth it, you can buy me a coke in return. Provided as is without any kind
  * of warranty. No attribution required.                  - Gregor Santner
  *
- * License: Creative Commons Zero (CC0 1.0)
+ * License of this file: Creative Commons Zero (CC0 1.0)
  *  http://creativecommons.org/publicdomain/zero/1.0/
  * ----------------------------------------------------------------------------
  */
@@ -22,9 +22,9 @@
  *
  * FILTER_ANDROID_TEXTVIEW output is intended to be used at simple Android TextViews,
  * were a limited set of html tags is supported. This allow to still display e.g. a simple
- * CHANGELOG.md file without inlcuding a WebView for showing HTML, or other additional UI-libraries.
+ * CHANGELOG.md file without including a WebView for showing HTML, or other additional UI-libraries.
  *
- * FILTER_HTMLPART is intended to be used at engines understanding most common HTML tags.
+ * FILTER_WEB is intended to be used at engines understanding most common HTML tags.
  */
 
 package io.github.gsantner.opoc.util;
@@ -38,39 +38,8 @@ import java.io.InputStreamReader;
 /**
  * Simple Markdown Parser
  */
-@SuppressWarnings({"WeakerAccess", "CaughtExceptionImmediatelyRethrown"})
+@SuppressWarnings({"WeakerAccess", "CaughtExceptionImmediatelyRethrown", "SameParameterValue", "unused", "SpellCheckingInspection", "RepeatedSpace", "SingleCharAlternation"})
 public class SimpleMarkdownParser {
-
-    public interface SimpleLineFilter {
-        String filterLine(String line);
-    }
-
-    public static final SimpleLineFilter FILTER_ANDROID_TEXTVIEW = new SimpleLineFilter() {
-        @Override
-        public String filterLine(String line) {
-            // TextView supports a limited set of html tags, most notably
-            // a href, b, big, font size&color, i, li, small, u
-            line = line
-                    .replace("~°", "&nbsp;&nbsp;") // double space/half tab
-                    .replaceAll("^### ([^<]*)", "<br/><big><b><font color='#000000'>$1</font></b></big>  ") // h3
-                    .replaceAll("^## ([^<]*)", "<br/><big><big><b><font color='#000000'>$1</font></b></big></big><br/>  ") // h2 (DEP: h3)
-                    .replaceAll("^# ([^<]*)", "<br/><big><big><big><b><font color='#000000'>$1</font></b></big></big></big><br/>  ") // h1 (DEP: h2,h3)
-                    .replaceAll("!\\[(.*?)\\]\\((.*?)\\)", "<a href=\\'$2\\'>$1</a>") // img
-                    .replaceAll("\\[(.*?)\\]\\((.*?)\\)", "<a href=\\'$2\\'>$1</a>") // a href (DEP: img)
-                    .replaceAll("<(http|https):\\/\\/(.*)>", "<a href='$1://$2'>$1://$2</a>") // a href (DEP: img)
-                    .replaceAll("^(-|\\*) ([^<]*)", "<font color='#000001'>&#8226;</font> $2  ") // unordered list + end line
-                    .replaceAll("^  (-|\\*) ([^<]*)", "&nbsp;&nbsp;<font color='#000001'>&#8226;</font> $2  ") // unordered list2 + end line
-                    .replaceAll("`([^<]*)`", "<font face='monospace'>$1</font>") // code
-                    .replace("\\*", "●") // temporary replace escaped star symbol
-                    .replaceAll("\\*\\*([^<]*)\\*\\*", "<b>$1</b>") // bold (DEP: temp star)
-                    .replaceAll("\\*([^<]*)\\*", "<i>$1</i>") // italic (DEP: temp star code)
-                    .replace("●", "*") // restore escaped star symbol (DEP: b,i)
-                    .replaceAll("  $", "<br/>") // new line (DEP: ul)
-            ;
-            return line.isEmpty() ? line + "<br/>" : line;
-        }
-    };
-
     private static SimpleMarkdownParser instance;
 
     public static SimpleMarkdownParser get() {
@@ -80,40 +49,94 @@ public class SimpleMarkdownParser {
         return instance;
     }
 
-    public static final SimpleLineFilter FILTER_HTMLPART = new SimpleLineFilter() {
+    public interface SmpFilter {
+        String filter(String text);
+    }
+
+    public final static SmpFilter FILTER_ANDROID_TEXTVIEW = new SmpFilter() {
         @Override
-        public String filterLine(String line) {
-            line = line
+        public String filter(String text) {
+            // TextView supports a limited set of html tags, most notably
+            // a href, b, big, font size&color, i, li, small, u
+
+            // Don't start new line if 2 empty lines and heading
+            while (text.contains("\n\n#")) {
+                text = text.replace("\n\n#", "\n#");
+            }
+
+            return text
+                    .replaceAll("(?s)<!--.*?-->", "")  // HTML comments
+                    .replace("\n\n", "\n<br/>\n") // Start new line if 2 empty lines
+                    .replace("~°", "&nbsp;&nbsp;") // double space/half tab
+                    .replaceAll("(?m)^### (.*)$", "<br/><big><b><font color='#000000'>$1</font></b></big><br/>") // h3
+                    .replaceAll("(?m)^## (.*)$", "<br/><big><big><b><font color='#000000'>$1</font></b></big></big><br/><br/>") // h2 (DEP: h3)
+                    .replaceAll("(?m)^# (.*)$", "<br/><big><big><big><b><font color='#000000'>$1</font></b></big></big></big><br/><br/>") // h1 (DEP: h2,h3)
+                    .replaceAll("!\\[(.*?)\\]\\((.*?)\\)", "<a href=\\'$2\\'>$1</a>") // img
+                    .replaceAll("\\[(.*?)\\]\\((.*?)\\)", "<a href=\\'$2\\'>$1</a>") // a href (DEP: img)
+                    .replaceAll("<(http|https):\\/\\/(.*)>", "<a href='$1://$2'>$1://$2</a>") // a href (DEP: img)
+                    .replaceAll("(?m)^([-*] )(.*)$", "<font color='#000001'>&#8226;</font> $2<br/>") // unordered list + end line
+                    .replaceAll("(?m)^  (-|\\*) ([^<]*)$", "&nbsp;&nbsp;<font color='#000001'>&#8226;</font> $2<br/>") // unordered list2 + end line
+                    .replaceAll("`([^<]*)`", "<font face='monospace'>$1</font>") // code
+                    .replace("\\*", "●") // temporary replace escaped star symbol
+                    .replaceAll("(?m)\\*\\*(.*)\\*\\*", "<b>$1</b>") // bold (DEP: temp star)
+                    .replaceAll("(?m)\\*(.*)\\*", "<i>$1</i>") // italic (DEP: temp star code)
+                    .replace("●", "*") // restore escaped star symbol (DEP: b,i)
+                    .replaceAll("(?m)  $", "<br/>") // new line (DEP: ul)
+                    ;
+        }
+    };
+
+    public final static SmpFilter FILTER_WEB = new SmpFilter() {
+        @Override
+        public String filter(String text) {
+            // Don't start new line if 2 empty lines and heading
+            while (text.contains("\n\n#")) {
+                text = text.replace("\n\n#", "\n#");
+            }
+
+            text = text
+                    .replaceAll("(?s)<!--.*?-->", "")  // HTML comments
+                    .replace("\n\n", "\n<br/>\n") // Start new line if 2 empty lines
                     .replaceAll("~°", "&nbsp;&nbsp;") // double space/half tab
-                    .replaceAll("^### ([^<]*)", "<h3>$1</h3>") // h3
-                    .replaceAll("^## ([^<]*)", "<h2>$1</h2>") /// h2 (DEP: h3)
-                    .replaceAll("^# ([^<]*)", "<h1>$1</h1>") // h1 (DEP: h2,h3)
+                    .replaceAll("(?m)^### (.*)$", "<h3>$1</h3>") // h3
+                    .replaceAll("(?m)^## (.*)$", "<h2>$1</h2>") /// h2 (DEP: h3)
+                    .replaceAll("(?m)^# (.*)$", "<h1>$1</h1>") // h1 (DEP: h2,h3)
                     .replaceAll("!\\[(.*?)\\]\\((.*?)\\)", "<img src=\\'$2\\' alt='$1' />") // img
                     .replaceAll("<(http|https):\\/\\/(.*)>", "<a href='$1://$2'>$1://$2</a>") // a href (DEP: img)
                     .replaceAll("\\[(.*?)\\]\\((.*?)\\)", "<a href=\\'$2\\'>$1</a>") // a href (DEP: img)
-                    .replaceAll("^(-|\\*) ([^<]*)", "<font color='#000001'>&#8226;</font> $2  ") // unordered list + end line
-                    .replaceAll("^  (-|\\*) ([^<]*)", "&nbsp;&nbsp;<font color='#000001'>&#8226;</font> $2  ") // unordered list2 + end line
+                    .replaceAll("(?m)^([-*] )(.*)$", "<font color='#000001'>&#8226;</font> $2  ") // unordered list + end line
+                    .replaceAll("(?m)^  (-|\\*) ([^<]*)$", "&nbsp;&nbsp;<font color='#000001'>&#8226;</font> $2  ") // unordered list2 + end line
                     .replaceAll("`([^<]*)`", "<code>$1</code>") // code
                     .replace("\\*", "●") // temporary replace escaped star symbol
-                    .replaceAll("\\*\\*([^<]*)\\*\\*", "<b>$1</b>") // bold (DEP: temp star)
-                    .replaceAll("\\*([^<]*)\\*", "<b>$1</b>") // italic (DEP: temp star)
+                    .replaceAll("(?m)\\*\\*(.*)\\*\\*", "<b>$1</b>") // bold (DEP: temp star)
+                    .replaceAll("(?m)\\*(.*)\\*", "<i>$1</i>") // italic (DEP: temp star code)
                     .replace("●", "*") // restore escaped star symbol (DEP: b,i)
-                    .replaceAll("  $", "<br/>") // new line (DEP: ul)
+                    .replaceAll("(?m)  $", "<br/>") // new line (DEP: ul)
             ;
-            return line.isEmpty() ? line + "<br/>" : line;
+            return text;
         }
     };
 
     //########################
     //##     Members
     //########################
+    private SmpFilter defaultSmpFilter;
     private String html;
 
-    public SimpleMarkdownParser parse(String filepath, SimpleLineFilter simpleLineFilter) throws IOException {
-        return parse(new FileInputStream(filepath), simpleLineFilter, "");
+    public SimpleMarkdownParser() {
+        setDefaultSmpFilter(FILTER_WEB);
     }
 
-    public SimpleMarkdownParser parse(InputStream inputStream, SimpleLineFilter simpleLineFilter, String lineMdPrefix) throws IOException {
+    public SimpleMarkdownParser setDefaultSmpFilter(SmpFilter defaultSmpFilter) {
+        this.defaultSmpFilter = defaultSmpFilter;
+        return this;
+    }
+
+    public SimpleMarkdownParser parse(String filepath, SmpFilter... smpFilters) throws IOException {
+        return parse(new FileInputStream(filepath), "", smpFilters);
+    }
+
+    public SimpleMarkdownParser parse(InputStream inputStream, String lineMdPrefix, SmpFilter... smpFilters) throws IOException {
         StringBuilder sb = new StringBuilder();
         BufferedReader br = null;
         String line;
@@ -121,7 +144,8 @@ public class SimpleMarkdownParser {
         try {
             br = new BufferedReader(new InputStreamReader(inputStream));
             while ((line = br.readLine()) != null) {
-                sb.append(simpleLineFilter.filterLine(lineMdPrefix + line));
+                sb.append(lineMdPrefix);
+                sb.append(line);
                 sb.append("\n");
             }
         } catch (IOException rethrow) {
@@ -135,19 +159,18 @@ public class SimpleMarkdownParser {
                 }
             }
         }
-        html = sb.toString().trim();
+        html = parse(sb.toString(), "", smpFilters).getHtml();
         return this;
     }
 
-    public SimpleMarkdownParser parse(String markdown, SimpleLineFilter simpleLineFilter, String lineMdPrefix) throws IOException {
-        StringBuilder sb = new StringBuilder();
-
-        for (String line : markdown.split("\n")) {
-            sb.append(simpleLineFilter.filterLine(lineMdPrefix + line));
-            sb.append("\n");
+    public SimpleMarkdownParser parse(String markdown, String lineMdPrefix, SmpFilter... smpFilters) throws IOException {
+        html = markdown;
+        if (smpFilters.length == 0) {
+            smpFilters = new SmpFilter[]{defaultSmpFilter};
         }
-
-        html = sb.toString().trim();
+        for (SmpFilter smpFilter : smpFilters) {
+            html = smpFilter.filter(html).trim();
+        }
         return this;
     }